Moved JVM startup code to separate folder
