+++ java/jvm/jni_helpers.hh
